refactor(fwa): split war/sync match-state helpers out of Fwa command
- add `src/commands/fwa/matchState.ts` for war/sync display, mismatch warning, points-site sync, and missed-sync helper logic
- move `WarStateForSync` and related utility functions out of `src/commands/Fwa.ts`
- keep `isMissedSyncClanForTest` exported from `Fwa.ts` via re-export to avoid test/API breakage
- reduce `Fwa.ts` size by removing duplicated in-file helper implementations
