+++ src/commands/Fwa.ts
@@ -22,12 +22,30 @@ import { prisma } from "../prisma";
 import { CoCService } from "../services/CoCService";
 import { CommandPermissionService } from "../services/CommandPermissionService";
 import { SettingsService } from "../services/SettingsService";
+import {
+  buildOutcomeMismatchWarning,
+  buildPointsMismatchWarning,
+  buildPointsSyncStatusLine,
+  buildSyncMismatchWarning,
+  deriveWarState,
+  deriveOpponentBalanceFromPrimarySnapshot,
+  formatMatchTypeLabel,
+  formatWarStateLabel,
+  getCurrentSyncFromPrevious,
+  getSyncDisplay,
+  getWarStateRemaining,
+  isMissedSyncClan,
+  isMissedSyncClanForTest,
+  isPointsSiteUpdatedForOpponent,
+  type WarStateForSync,
+  withSyncModeLabel,
+} from "./fwa/matchState";
+export { isMissedSyncClanForTest } from "./fwa/matchState";
 
 const POINTS_BASE_URL = "https://points.fwafarm.com/clan?tag=";
 const TIEBREAK_ORDER = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 const CACHE_REFRESH_DELAY_MS = 30 * 60 * 1000;
 const WAR_END_RECHECK_MS = 10 * 60 * 1000;
-const MISSED_SYNC_WINDOW_MS = 2 * 60 * 60 * 1000;
 const DISCORD_CONTENT_MAX = 2000;
 const POINTS_CACHE_VERSION = 5;
 const MATCHUP_CACHE_VERSION = 5;
@@ -79,7 +97,6 @@ type MatchupCacheEntry = {
 };
 
 const PREVIOUS_SYNC_KEY = "previousSyncNum";
-type WarStateForSync = "preparation" | "inWar" | "notInWar";
 
 function normalizeTag(input: string): string {
   return input.trim().toUpperCase().replace(/^#/, "");
@@ -1124,193 +1141,6 @@ async function getSourceOfTruthSync(settings: SettingsService): Promise<number |
   const parsed = Number(raw);
   if (!Number.isFinite(parsed)) return null;
   return Math.trunc(parsed);
-}
-
-function deriveWarState(rawState: string | null | undefined): WarStateForSync {
-  const state = String(rawState ?? "").toLowerCase();
-  if (state.includes("preparation")) return "preparation";
-  if (state.includes("inwar")) return "inWar";
-  return "notInWar";
-}
-
-function getCurrentSyncFromPrevious(
-  previousSync: number | null,
-  warState: WarStateForSync
-): number | null {
-  if (previousSync === null) return null;
-  if (warState === "notInWar") return null;
-  return previousSync + 1;
-}
-
-function getSyncDisplay(
-  previousSync: number | null,
-  warState: WarStateForSync
-): string {
-  if (previousSync === null) return "unknown";
-  const current = previousSync + 1;
-  if (warState === "notInWar") {
-    return `between #${previousSync} and #${current}`;
-  }
-  return `#${current}`;
-}
-
-function getSyncModeFromPrevious(previousSync: number | null): "high" | "low" | null {
-  if (previousSync === null) return null;
-  return (previousSync + 1) % 2 === 0 ? "high" : "low";
-}
-
-function withSyncModeLabel(syncText: string, previousSync: number | null): string {
-  const mode = getSyncModeFromPrevious(previousSync);
-  if (!mode) return syncText;
-  return `${syncText} (${mode === "high" ? "High Sync" : "Low Sync"})`;
-}
-
-function formatWarStateLabel(warState: WarStateForSync): string {
-  if (warState === "preparation") return "preparation";
-  if (warState === "inWar") return "battle day";
-  return "no war";
-}
-
-function formatMatchTypeLabel(
-  matchType: "FWA" | "BL" | "MM" | "UNKNOWN",
-  inferred: boolean
-): string {
-  if (!inferred) return matchType;
-  return `${matchType} \u26A0\uFE0F`;
-}
-
-function isPointsSiteUpdatedForOpponent(
-  primary: PointsSnapshot,
-  opponentTag: string,
-  previousSync: number | null
-): boolean {
-  const normalizedOpponent = normalizeTag(opponentTag);
-  const hasOpponent = primary.winnerBoxTags.map((t) => normalizeTag(t)).includes(normalizedOpponent);
-  if (!hasOpponent) return false;
-  if (
-    previousSync !== null &&
-    primary.winnerBoxSync !== null &&
-    Number.isFinite(primary.winnerBoxSync) &&
-    primary.winnerBoxSync <= previousSync
-  ) {
-    return false;
-  }
-  return true;
-}
-
-function deriveOpponentBalanceFromPrimarySnapshot(
-  primary: PointsSnapshot,
-  primaryTag: string,
-  opponentTag: string
-): number | null {
-  const normalizedPrimary = normalizeTag(primaryTag);
-  const normalizedOpponent = normalizeTag(opponentTag);
-  if (
-    primary.headerPrimaryTag === normalizedPrimary &&
-    primary.headerOpponentTag === normalizedOpponent
-  ) {
-    return primary.headerOpponentBalance;
-  }
-  if (
-    primary.headerPrimaryTag === normalizedOpponent &&
-    primary.headerOpponentTag === normalizedPrimary
-  ) {
-    return primary.headerPrimaryBalance;
-  }
-  return null;
-}
-
-function buildPointsMismatchWarning(
-  label: string,
-  expected: number | null | undefined,
-  actual: number | null | undefined
-): string | null {
-  if (
-    expected === null ||
-    expected === undefined ||
-    actual === null ||
-    actual === undefined ||
-    Number.isNaN(expected) ||
-    Number.isNaN(actual)
-  ) {
-    return null;
-  }
-  if (expected === actual) return null;
-  return `\u26A0\uFE0F ${label} points mismatch: expected ${expected}, site ${actual}.`;
-}
-
-function buildSyncMismatchWarning(
-  expectedSync: number | null | undefined,
-  siteSync: number | null | undefined
-): string | null {
-  if (
-    expectedSync === null ||
-    expectedSync === undefined ||
-    siteSync === null ||
-    siteSync === undefined ||
-    Number.isNaN(expectedSync) ||
-    Number.isNaN(siteSync)
-  ) {
-    return null;
-  }
-  if (expectedSync === siteSync) return null;
-  return `\u26A0\uFE0F Sync # mismatch: expected #${expectedSync}, site #${siteSync}.`;
-}
-
-function buildOutcomeMismatchWarning(
-  expectedOutcome: "WIN" | "LOSE" | null | undefined,
-  siteOutcome: "WIN" | "LOSE" | null | undefined
-): string | null {
-  if (!siteOutcome) return null;
-  if (expectedOutcome === siteOutcome) return null;
-  return `\u26A0\uFE0F Outcome mismatch: expected ${expectedOutcome ?? "UNKNOWN"}, site ${siteOutcome}.`;
-}
-
-function buildPointsSyncStatusLine(siteUpdated: boolean, hasMismatch: boolean): string {
-  if (!siteUpdated) {
-    return ":hourglass_flowing_sand: points.fwafarm is not updated for this matchup yet.";
-  }
-  if (hasMismatch) {
-    return ":broken_chain: out of sync with points site";
-  }
-  return ":white_check_mark: data in sync with points.fwafarm";
-}
-
-function getWarStateRemaining(
-  war: { startTime?: string | null; endTime?: string | null } | null | undefined,
-  warState: WarStateForSync
-): string {
-  if (warState === "notInWar") return "n/a";
-  const startMs = parseCocApiTime(war?.startTime);
-  const endMs = parseCocApiTime(war?.endTime);
-  const targetMs = warState === "preparation" ? startMs : endMs;
-  if (targetMs === null || !Number.isFinite(targetMs)) return "unknown";
-  return `<t:${Math.floor(targetMs / 1000)}:R>`;
-}
-
-function isMissedSyncClan(input: {
-  baselineWarStartMs: number | null;
-  clanWarState: WarStateForSync;
-  clanWarStartMs: number | null;
-  nowMs: number;
-}): boolean {
-  const { baselineWarStartMs, clanWarState, clanWarStartMs, nowMs } = input;
-  if (baselineWarStartMs === null || !Number.isFinite(baselineWarStartMs)) return false;
-  const deadlineMs = baselineWarStartMs + MISSED_SYNC_WINDOW_MS;
-  if (clanWarState === "notInWar") {
-    return nowMs >= deadlineMs;
-  }
-  if (clanWarStartMs === null || !Number.isFinite(clanWarStartMs)) return false;
-  return clanWarStartMs > deadlineMs;
-}
-
-export function isMissedSyncClanForTest(input: {
-  baselineWarStartMs: number | null;
-  clanWarState: WarStateForSync;
-  clanWarStartMs: number | null;
-  nowMs: number;
-}): boolean {
-  return isMissedSyncClan(input);
 }
 
 async function resolveMatchTypeWithFallback(params: {
